// frontend/src/api-client/model/image-response-dto.ts
/**
 * Inkweld API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageDataDto } from './image-data-dto';
import { ImageUsageDto } from './image-usage-dto';


export interface ImageResponseDto { 
    /**
     * Timestamp of creation (Unix seconds)
     */
    created: number;
    /**
     * List of generated image data objects
     */
    data: Array<ImageDataDto>;
    /**
     * Token usage information (gpt-image-1 only)
     */
    usage?: ImageUsageDto;
    /**
     * Source of the image generation (e.g., \"openai\")
     */
    source: string;
}